# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "vms Api Documentation",
    "description": "vms Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "vms Logo",
      "altText": "vms Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "vms API Support",
      "url": "https://vms.com/info/support",
      "email": "support@vms.com"
    },
    "termsOfService": "https://vms.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/kilometrage": {
      "get": {
        "summary": "List Kilometrage",
        "description": "Kilometrage",
        "operationId": "ListKilometrage",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Kilometrage"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListKilometrage"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/kilometrage/view/{id}": {
      "get": {
        "summary": "View Kilometrage",
        "description": "Vue Kilometrage",
        "operationId": "ViewKilometrage",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Kilometrage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewKilometrage"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/kilometrage/add": {
      "post": {
        "summary": "Add Kilometrage",
        "description": "Ajouter un nouveau",
        "operationId": "AddKilometrage",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Kilometrage"
        ],
        
        "requestBody": {
          "description": "Add Kilometrage",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddKilometrage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewKilometrage"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/kilometrage/edit/{id}": {
      "put": {
        "summary": "Edit Kilometrage",
        "description": "modifier",
        "operationId": "EditKilometrage",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Kilometrage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Kilometrage",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditKilometrage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditKilometrage"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
    "/kilometrage/delete/{id}": {
      "delete": {
        "summary": "Delete Kilometrage",
        "description": "Kilometrage",
        "operationId": "DeleteKilometrage",
        "security": [
          {
            "ApiKeyBearer": []
          }
        ],
        "tags": [
          "Kilometrage"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "UnAuthenticated request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnAuthenticated"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "ApiKeyBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ListKilometrage": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "compteur": {
            "description": "Compteur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "date_created": {
            "description": "Date Created",
            "type": "string"
          },
          "date_updated": {
            "description": "Date Updated",
            "type": "string"
          },
          "user_id": {
            "description": "User Id",
            "type": "string"
          }
        }
      },
      "ViewKilometrage": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "compteur": {
            "description": "Compteur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          },
          "date_created": {
            "description": "Date Created",
            "type": "string"
          },
          "date_updated": {
            "description": "Date Updated",
            "type": "string"
          },
          "user_id": {
            "description": "User Id",
            "type": "string"
          }
        }
      },
      "AddKilometrage": {
        "type": "object",
        "properties": {
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "compteur": {
            "description": "Compteur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          }
        }
      },
      "EditKilometrage": {
        "type": "object",
        "properties": {
          "vehicule_id": {
            "description": "Vehicule Id",
            "type": "string"
          },
          "date": {
            "description": "Date",
            "type": "string"
          },
          "compteur": {
            "description": "Compteur",
            "type": "string"
          },
          "note": {
            "description": "Note",
            "type": "string"
          },
          "photo": {
            "description": "Photo",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
...
      "UnAuthenticated": {
        "type": "object",
        "properties": {
          "code": {
            "description": "401",
            "type": "integer"
          },
          "message": {
            "description": "UnAuthenticated",
            "type": "string"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}